fix(test): isolate the test to avoid affect others
