+++ run_test.go
@@ -1,18 +1,24 @@
 package playwright
 
 import (
+	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"os"
+	"path/filepath"
+	"runtime"
 	"strings"
 	"testing"
+
+	"github.com/mitchellh/go-ps"
+	"github.com/stretchr/testify/require"
 )
 
 func TestDriverInstall(t *testing.T) {
 	driverPath := t.TempDir()
 	driver, err := NewDriver(&RunOptions{
 		DriverDirectory: driverPath,
-		Browsers:        []string{"firefox"},
+		Browsers:        []string{getBrowserName()},
 		Verbose:         true})
 	if err != nil {
 		t.Fatalf("could not start driver: %v", err)
@@ -58,4 +64,78 @@ func TestDriverDownloadHostEnv(t *testing.T) {
 	if err == nil || !strings.Contains(err.Error(), "404 Not Found") || !strings.Contains(uri, "/builds/driver") {
 		t.Fatalf("PLAYWRIGHT_DOWNLOAD_HOST do not work: %v", err)
 	}
+}
+
+func TestShouldNotHangWhenPlaywrightUnexpectedExit(t *testing.T) {
+	if runtime.GOOS != "windows" || getBrowserName() != "chromium" {
+		t.Skip("chromium on windows only")
+		return
+	}
+
+	pw, err := Run()
+	require.NoError(t, err)
+	defer func() {
+		_ = pw.Stop()
+	}()
+	browser, err := pw.Chromium.Launch()
+	require.NoError(t, err)
+	context, err := browser.NewContext()
+	require.NoError(t, err)
+
+	err = killPlaywrightProcess()
+	require.NoError(t, err)
+
+	_, err = context.NewPage()
+	require.Error(t, err)
+}
+
+// find and kill playwright process, only for Windows
+func killPlaywrightProcess() error {
+	all, err := ps.Processes()
+	if err != nil {
+		return err
+	}
+	for _, process := range all {
+		if process.Executable() == "node" || process.Executable() == "node.exe" {
+			parent, err := ps.FindProcess(process.PPid())
+			if err != nil {
+				return err
+			}
+			if parent.Executable() == "cmd.exe" {
+				grandpa, err := ps.FindProcess(parent.PPid())
+				if err != nil {
+					return err
+				}
+				if strings.HasPrefix(grandpa.Executable(), "__debug_bin") || grandpa.Executable() == filepath.Base(os.Args[0]) {
+					if err := killProcessByPid(parent.Pid()); err != nil {
+						return err
+					}
+					if err := killProcessByPid(process.Pid()); err != nil {
+						return err
+					}
+					return nil
+				}
+			}
+		}
+	}
+	return fmt.Errorf("playwright process not found")
+}
+
+func killProcessByPid(pid int) error {
+	process, err := os.FindProcess(pid)
+	if err != nil {
+		return err
+	}
+	if err := process.Kill(); err != nil {
+		return err
+	}
+	return nil
+}
+
+func getBrowserName() string {
+	browserName, hasEnv := os.LookupEnv("BROWSER")
+	if hasEnv {
+		return browserName
+	}
+	return "chromium"
 }
